Reorganize files into new directories
