+++ src/components/Layout/index.tsx
@@ -1,5 +1,5 @@
-import Header from "./Header";
-import Footer from "./Footer";
+import Header from "./Header.tsx";
+import Footer from "./Footer.tsx";
 import { Outlet } from "react-router";
 
 export default function Layout() {
