+++ src/main.tsx
@@ -2,9 +2,9 @@ import { StrictMode } from "react";
 import { createRoot } from "react-dom/client";
 import { BrowserRouter, Routes, Route } from "react-router";
 import "./index.css";
-import Layout from "./components/Layout.tsx";
-import Home from "./routes/Home.tsx";
-import About from "./routes/About.tsx";
+import Layout from "./components/Layout/index.tsx";
+import Home from "./routes/Home/index.tsx";
+import About from "./routes/About/index.tsx";
 
 const root = document.getElementById("root");
