--- src/routes/About.tsx
@@ -1,7 +1,0 @@
-export default function About() {
-  return (
-    <main className="min-h-[75vh] grid place-items-center bg-primary">
-      <p className="text-xl">About page</p>
-    </main>
-  );
-}
