+++ src/routes/About/index.tsx
@@ -1,0 +1,7 @@
+export default function About() {
+  return (
+    <main className="min-h-[25vh] grid place-items-center bg-primary">
+      <h3 className="text-3xl text-secondary">About Code Cafe</h3>
+    </main>
+  );
+}
