--- src/routes/Home.tsx
@@ -1,15 +1,0 @@
-import Hero from "../components/Hero";
-import Features from "../components/Features";
-import CTA from "../components/CTA";
-import Showcase from "../components/Showcase.tsx";
-
-export default function Home() {
-  return (
-    <>
-      <Hero />
-      <Features />
-      <Showcase />
-      <CTA />
-    </>
-  );
-}
