+++ src/routes/Home/index.tsx
@@ -1,0 +1,15 @@
+import Hero from "./Hero.tsx";
+import Features from "./Features.tsx";
+import CTA from "./CTA.tsx";
+import Showcase from "./Showcase.tsx";
+
+export default function Home() {
+  return (
+    <>
+      <Hero />
+      <Features />
+      <Showcase />
+      <CTA />
+    </>
+  );
+}
